Fix log paths and runtime env
