+++ pauta_aneel/keywords.example
@@ -1,0 +1,2 @@
+consulta publica
+leilao
