+++ requirements.txt
@@ -5,3 +5,5 @@ python-2captcha
 pillow
 pytesseract
 colorama
+requests
+beautifulsoup4
