+++ sorteio_aneel/sorteio_aneel.py
@@ -12,11 +12,16 @@ from email.mime.multipart import MIMEMultipart
 import unicodedata
 import tempfile
 import subprocess
-from urllib.parse import urljoin
-import hashlib
+import json
+
+# Diretório de dados e arquivos de log
+DATA_DIR = os.environ.get("SORTEIO_DATA_DIR", os.path.join(os.path.expanduser("~"), ".sorteio_aneel"))
+os.makedirs(DATA_DIR, exist_ok=True)
+LOG_FILE = os.environ.get("SORTEIO_LOG_FILE", os.path.join(DATA_DIR, "sorteio_aneel.log"))
+KEYWORDS_FILE = os.environ.get("KEYWORDS_FILE", os.path.join(DATA_DIR, ".sorteio_aneel_keywords"))
 
 # === Registro de data/hora de execução no log ===
-LOG_FILE = "/tmp/pauta_aneel_cron.log"
+
 def registrar_log(mensagem):
     try:
         with open(LOG_FILE, "a") as f:
@@ -33,13 +38,17 @@ SMTP_SERVER = os.environ.get("SMTP_SERVER", "smtp.hscl.adv.br")
 SMTP_PORT = int(os.environ.get("SMTP_PORT", "587"))
 SMTP_USER = os.environ.get("SMTP_USER", "ro-dou@hscl.adv.br")
 SMTP_PASSWORD = os.environ.get("SMTP_PASSWORD", "Aasn1989")
-EMAIL_TO = os.environ.get("EMAIL_TO", "asamia@isacteep.com.br,ary@hscl.adv.br")
+EMAIL_TO = os.environ.get(
+    "EMAIL_TO",
+    "ARY@HSCL.ADV.BR,asamia@isacteep.com.br"
+)
 
-BASE_URL = "https://www2.aneel.gov.br/aplicacoes_liferay/noticias_area/?idAreaNoticia=425"
+BASE_URL = "https://www2.aneel.gov.br/aplicacoes_liferay/noticias_area/?idAreaNoticia=424"
 SITE_PREFIX = "https://www2.aneel.gov.br"
 
-HASH_RESULT_FILE = os.environ.get("HASH_RESULT_FILE", "/tmp/ultimo_resultado_aneel_425.txt")
-KEYWORDS_FILE = ".pauta_aneel_keywords"
+LAST_RESULT_FILE = os.environ.get(
+    "LAST_RESULT_FILE", os.path.join(DATA_DIR, "ultimo_resultado_aneel.json")
+)
 
 # Função para carregar termos de pesquisa do arquivo externo, se existir
 def load_keywords():
@@ -53,7 +62,7 @@ def load_keywords():
             registrar_log(f"Erro ao ler {KEYWORDS_FILE}: {e}")
     # fallback para lista padrão
     return [
-        "consulta publica", "tomada de subsidio", "consulta externa", "audiencia publica",
+        "consulta publica", "reajuste tarifario", "rbse", "tomada de subsidio", "consulta externa", "audiencia publica",
         "leilao", "isa energia", "cteep", "companhia de transmissao de energia eletrica paulista",
         "interligacao eletrica"
     ]
@@ -80,11 +89,27 @@ def parse_date(date_str):
     except Exception:
         return None
 
+def ler_ultimo_resultado():
+    try:
+        with open(LAST_RESULT_FILE, "r") as f:
+            return json.load(f)
+    except Exception:
+        return {"data_encontrada": None, "items": []}
+
+def salvar_ultimo_resultado(data_encontrada, items):
+    try:
+        with open(LAST_RESULT_FILE, "w") as f:
+            json.dump({
+                "data_encontrada": data_encontrada,
+                "items": items
+            }, f)
+    except Exception as e:
+        registrar_log(f"Erro ao salvar último resultado: {e}")
+
 def find_nearest_date_link(target_date=None):
     headers = {"User-Agent": "Mozilla/5.0"}
     response = requests.get(BASE_URL, timeout=30, headers=headers)
     if not response.ok:
-        print(f"Erro ao acessar ANEEL: {response.status_code}")
         registrar_log(f"Erro ao acessar ANEEL: {response.status_code}")
         return None, None, None
     soup = BeautifulSoup(response.text, "html.parser")
@@ -109,7 +134,6 @@ def find_nearest_date_link(target_date=None):
         if isinstance(target_date, str):
             target_date = parse_date(target_date)
         if not target_date:
-            print("Data inválida informada.")
             registrar_log("Data inválida informada.")
             return None, None, None
     for date_found, link_found, link_text in found_rows:
@@ -123,87 +147,57 @@ def extract_items_from_tr(url):
     response = requests.get(url, timeout=30, headers=headers)
     response.encoding = response.apparent_encoding
     if not response.ok:
-        print(f"Erro ao acessar documento: {response.status_code}")
         registrar_log(f"Erro ao acessar documento: {response.status_code}")
         return []
     soup = BeautifulSoup(response.text, "html.parser")
-
-    if soup.body:
-        content_elements = []
-        for tag in soup.body.descendants:
-            if tag.name in ['p', 'div', 'li']:
-                content_elements.append(tag)
-    else:
-        content_elements = soup.find_all(['p', 'div', 'li'])
-
-    idx = 0
+    all_tr = soup.find_all("tr")
+    main_tr = max(all_tr, key=lambda tr: len(tr.get_text()), default=None)
+    if not main_tr:
+        registrar_log("Nenhum <tr> relevante encontrado.")
+        return []
+    full_text = main_tr.get_text(separator=" ", strip=True)
+    marker_regex = re.compile(r'(?<!\d)(\d{1,3})\. Processo: \d{5,}\.')
+    matches = list(marker_regex.finditer(full_text))
+    items = []
+    for idx, match in enumerate(matches):
+        start = match.start()
+        end = matches[idx+1].start() if idx+1 < len(matches) else len(full_text)
+        item = full_text[start:end].strip()
+        items.append(item)
     final_results = []
-    while idx < len(content_elements):
-        el = content_elements[idx]
-        texto = el.get_text(separator=" ", strip=True)
-        m_proc = re.match(r'^\d+\.\s*Processo\s*:\s*[\d\.]+/\d{4}-\d{2}', texto)
-        if m_proc:
-            bloco_texto = texto
-            bloco_idx_start = idx
-            processo_numero = None
-            processo_numero_pdf = ""
-            m_proc_n = re.search(r'Processo\s*:\s*([\d\.]+/\d{4}-\d{2})', bloco_texto)
-            if m_proc_n:
-                processo_numero = m_proc_n.group(1)
-                processo_numero_pdf = re.sub(r'[\./-]', '', processo_numero)
-            pdfs = []
-            doc_count = 0
-            idx_next = idx + 1
-            bloco_extra_text = ""
-            while idx_next < len(content_elements):
-                texto_next = content_elements[idx_next].get_text(separator=" ", strip=True)
-                if re.match(r'^\d+\.\s*Processo\s*:\s*[\d\.]+/\d{4}-\d{2}', texto_next):
-                    break
-                for a in content_elements[idx_next].find_all('a', href=True):
-                    href = a['href']
-                    if 'pdf' in href.lower():
-                        doc_count += 1
-                        pdf_url = urljoin(url, href)
-                        doc_index = f" - doc{doc_count}" if doc_count > 1 else ""
-                        pdf_filename = f"Minutas de voto e ato - {processo_numero_pdf}{doc_index}.pdf"
-                        pdfs.append({"pdf_url": pdf_url, "pdf_filename": pdf_filename})
-                bloco_extra_text += " " + texto_next
-                idx_next += 1
-            bloco_texto_full = bloco_texto + bloco_extra_text
-            if palavra_chave_no_texto(bloco_texto_full, KEYWORDS):
-                final_results.append({
-                    "text": bloco_texto_full.strip(),
-                    "processo_numero": processo_numero,
-                    "processo_numero_pdf": processo_numero_pdf,
-                    "pdfs": pdfs
-                })
-            idx = idx_next
-        else:
-            idx += 1
+    for item in items:
+        if palavra_chave_no_texto(item, KEYWORDS):
+            item_clean = re.sub(r'\s+', ' ', item).strip()
+            final_results.append(item_clean)
     return final_results
 
 def gerar_pdf_da_pagina(url, pdf_file):
     try:
-        # Gera o PDF diretamente da URL, ignorando erros de carregamento de recursos externos
+        headers = {'User-Agent': 'Mozilla/5.0'}
+        html = requests.get(url, headers=headers, timeout=30).text
+        with tempfile.NamedTemporaryFile(delete=False, suffix=".html") as tmp_html:
+            tmp_html.write(html.encode("utf-8"))
+            html_path = tmp_html.name
         result = subprocess.run(
-            ["wkhtmltopdf", "--quiet", "--load-error-handling", "ignore", url, pdf_file],
+            ["wkhtmltopdf", "--quiet", html_path, pdf_file],
             stdout=subprocess.PIPE,
             stderr=subprocess.PIPE
         )
         if result.returncode != 0:
-            print("Erro ao gerar PDF (wkhtmltopdf):", result.stderr.decode())
             registrar_log(f"Erro ao gerar PDF (wkhtmltopdf): {result.stderr.decode()}")
+            os.remove(html_path)
             return False
+        os.remove(html_path)
         return True
     except Exception as e:
-        print(f"Erro ao gerar PDF: {e}")
         registrar_log(f"Erro ao gerar PDF: {e}")
         return False
 
 def send_email(subject, body, pdf_path=None):
     msg = MIMEMultipart()
     msg["From"] = SMTP_USER
-    msg["To"] = EMAIL_TO
+    destinatarios = [e.strip() for e in EMAIL_TO.replace(';', ',').split(',') if e.strip()]
+    msg["To"] = ", ".join(destinatarios)
     msg["Subject"] = subject
     pdf_attached = False
     if pdf_path and os.path.exists(pdf_path) and os.path.getsize(pdf_path) > 0:
@@ -217,36 +211,20 @@ def send_email(subject, body, pdf_path=None):
             msg.attach(part)
             pdf_attached = True
     else:
-        print("PDF não gerado ou está vazio, não será anexado.")
+        registrar_log("PDF não gerado ou está vazio, não será anexado.")
         body += "\n\nATENÇÃO: Não foi possível anexar o PDF da página, pois ocorreu um erro na geração do arquivo.\n"
     msg.attach(MIMEText(body, "plain"))
 
     try:
         with smtplib.SMTP(SMTP_SERVER, SMTP_PORT) as server:
             server.login(SMTP_USER, SMTP_PASSWORD)
-            to_emails = [email.strip() for email in EMAIL_TO.split(",")]
-            server.sendmail(SMTP_USER, to_emails, msg.as_string())
+            server.sendmail(SMTP_USER, destinatarios, msg.as_string())
         print("E-mail enviado com sucesso.")
         registrar_log(f"E-mail enviado para {EMAIL_TO}")
-        registrar_log("Corpo do e-mail:\n" + body + "\n---")
+        registrar_log("Corpo do e-mail:\n" + body)
     except Exception as e:
         print(f"Erro ao enviar e-mail: {e}")
         registrar_log(f"Falha ao enviar e-mail: {e}")
-
-def hash_do_resultado(itens):
-    texto = "\n".join([item['text'] for item in itens])
-    return hashlib.sha256(texto.encode('utf-8')).hexdigest()
-
-def ler_ultimo_hash(arquivo):
-    try:
-        with open(arquivo, "r") as f:
-            return f.read().strip()
-    except FileNotFoundError:
-        return None
-
-def salvar_hash(arquivo, hash_str):
-    with open(arquivo, "w") as f:
-        f.write(hash_str)
 
 def main():
     if len(sys.argv) > 1:
@@ -255,17 +233,29 @@ def main():
     else:
         data_pesquisa = None
         forcar_envio = False
+
     hoje = datetime.now()
     hoje_str = hoje.strftime("%d/%m/%Y")
     url, link_text, data_encontrada = find_nearest_date_link(data_pesquisa or hoje_str)
     pdf_path = None
+
     if not url or not data_encontrada:
         print("Nenhum link associado à data encontrada.")
-        subject = f"{hoje_str} Busca Pauta ANEEL - Nenhuma data encontrada"
-        body = "Nao encontrada pauta para data indicada! Atenciosamente, Ary Abdo!"
+        subject = f"{hoje_str} Busca Sorteio ANEEL - Nenhuma data encontrada"
+        body = "Nao encontrado sorteio para data indicada! Atenciosamente, Ary Abdo!"
         if forcar_envio:
             send_email(subject, body)
         registrar_log("Nenhum link associado à data encontrada. Nenhum e-mail enviado.")
+        return
+
+    items = extract_items_from_tr(url)
+    ultimo_resultado = ler_ultimo_resultado()
+    items_anteriores = ultimo_resultado.get("items", [])
+    data_encontrada_anterior = ultimo_resultado.get("data_encontrada")
+
+    if not forcar_envio and items == items_anteriores:
+        print("Nenhuma atualização nos itens encontrados.")
+        registrar_log("Nenhuma atualização nos itens encontrados.")
         return
 
     if url:
@@ -275,40 +265,24 @@ def main():
         if not sucesso_pdf:
             print("PDF não gerado, mas tentando anexar se existir.")
 
-    items = extract_items_from_tr(url)
-    subject = f"{hoje_str} Busca Pauta ANEEL - {data_encontrada} - {link_text}"
-
+    subject = f"{hoje_str} Busca Sorteio ANEEL - {data_encontrada} - {link_text}"
     if items:
-        body = "Foram encontrados os processos listados abaixo na pauta da próxima reuniao da ANEEL:\n\n"
+        body = (
+            "Foram encontrados os processos listados abaixo no sorteio realizado pela ANEEL:\n\n"
+            + "\n\n".join(items)
+            + "\n\nAtenciosamente,\nAry Abdo"
+        )
         for item in items:
-            body += item["text"] + "\n"
-            registrar_log("Processo encontrado:\n" + item["text"])
-            if item["pdfs"]:
-                for pdf in item["pdfs"]:
-                    body += f"(Acesse o documento PDF: {pdf['pdf_url']})\n"
-                body += "\n"
-            else:
-                body += "Documentos nao disponibilizados.\n\n"
-        body += "\nAtenciosamente,\nAry Abdo"
+            registrar_log(f"Processo encontrado: {item}")
     else:
-        body = "Ola! Nao foram encontrados processos listados na pauta na data de pesquisa!\n\nAtenciosamente, Ary Abdo!"
+        body = "Ola! Nao foram encontrados processos sorteados na data de pesquisa!\n\nAtenciosamente, Ary Abdo!"
         registrar_log("Nenhum processo relevante encontrado.")
 
     print("Itens relevantes encontrados:" if items else "Nenhum item relevante encontrado.")
     print(body)
-
-    hash_arquivo = HASH_RESULT_FILE
-    hash_atual = hash_do_resultado(items)
-    hash_antigo = ler_ultimo_hash(hash_arquivo)
-
-    if forcar_envio or (hash_atual != hash_antigo):
-        send_email(subject, body, pdf_path)
-        salvar_hash(hash_arquivo, hash_atual)
-        registrar_log("Conteúdo alterado/enviado. Hash atualizado.")
-    else:
-        print("Nenhuma atualização no conteúdo da pauta ANEEL.")
-        registrar_log("Nenhuma atualização na data da pauta ANEEL.")
-
+    send_email(subject, body, pdf_path)
+    if not forcar_envio:
+        salvar_ultimo_resultado(data_encontrada, items)
     if pdf_path and os.path.exists(pdf_path):
         os.remove(pdf_path)
